[유빈] 로그인 구현 중 서버 작동 확인 후 ejs 파일 테스트 에러

+++ models/User.js
@@ -4,50 +4,46 @@ const {
 } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
   class USER extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
     static associate(models) {
       // define association here
     }
   }
+
   USER.init({
     token: {
       type: DataTypes.STRING(255),
-      defaultValue: null, 
+      defaultValue: null,
       unique: false
-    }, 
+    },
     id: {
-      primaryKey: true, 
+      primaryKey: true,
       type: DataTypes.STRING(15),
-      unique: false, 
+      unique: false,
       allowNull: false
-    }, 
+    },
     pw: {
-      type: DataTypes.STRING(255), 
+      type: DataTypes.STRING(255),
       allowNull: false
-    }, 
+    },
     salt: {
-      type: DataTypes.INTEGER, 
+      type: DataTypes.BIGINT,
       allowNull: false
     },
     nickname: {
-      type: DataTypes.STRING(15), 
-      allowNull: false, 
+      type: DataTypes.STRING(15),
+      allowNull: false,
       defaultValue: "unknown"
     }
-  }, 
+  },
   {
-    sequelize, 
-    timestamps: false, 
-    freezeTableName: true, 
+    sequelize,
+    timestamps: false,
+    freezeTableName: true,
     modelName: 'USER',
-    paranoid: false, 
+    paranoid: false,
     charset: 'utf8',
     collate: 'utf8_general_ci'
   });
   return USER;
-  
+
 };
